+++ src/index.js
@@ -1,0 +1,3 @@
+import { getLiteratureInfos } from "./UrlReading"
+
+getLiteratureInfos();
