+++ webpack.config.js
@@ -1,0 +1,19 @@
+const path = require('path');
+const CopyWebpackPlugin = require('copy-webpack-plugin');
+
+module.exports = (env) => ({
+  mode: env.mode,
+  entry: './src/index.js',
+  output: {
+    filename: 'main.js',
+    path: path.resolve(__dirname, 'dist'),
+  },
+  plugins: [
+    new CopyWebpackPlugin({
+      patterns: [
+        { from: 'src/manifest.json', to: 'manifest.json' },
+        { from: 'src/icon.png', to: 'icon.png' }
+      ],
+    }),
+  ],
+});
